requirements.txt
python-imagesearch-drov0
keyboard
ahk
pyautogui